pyproject.toml
[project]
name = "couchdb-jwt-proxy"
version = "1.0.0"
description = "HTTP proxy for CouchDB with JWT authentication"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Roady Team"}
]

dependencies = [
    "fastapi",
    "uvicorn",
    "pyjwt>=2.4.0", # Includes PyJWKClient for RS256 validation
    "cryptography", # For RSA key handling
    "httpx",
    "python-dotenv",
    "pydantic",
    "clerk-backend-api", # Clerk Backend API for session management
    "slowapi>=0.1.5", # Rate limiting library
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
pythonpath = ["src"]

[tool.pytest_asyncio]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["couchdb_jwt_proxy"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
